Moved Route Generation to route page | Removed wiki-generation page since it's no longer necessary
