+++ src/routes/game-routes/+page.svelte
@@ -2,7 +2,7 @@
 import BaseModal from "$lib/components/BaseModal.svelte";
 import Button from "$lib/components/Button.svelte";
 import TextInput from "$lib/components/TextInput.svelte";
-import { getToastStore, popup } from "@skeletonlabs/skeleton";
+import { Tab, TabGroup, getToastStore, popup } from "@skeletonlabs/skeleton";
 import { IconDotsVertical, IconTrash } from "@tabler/icons-svelte";
 import { invoke } from "@tauri-apps/api";
 import { selectedWiki } from "../../store";
@@ -24,6 +24,7 @@ let positionModalOpen: boolean = false;
 let newEncounterType: string = "";
 let routeBeingEdited: string = "";
 let loading = false;
+let tabSet = 0;
 
 async function createNewRoute() {
   $routes.routes[routeName.trim()] = {
@@ -103,6 +104,21 @@ async function updatePosition() {
     { dir: BaseDirectory.AppData },
   );
 }
+
+async function generateRoutePages() {
+  loading = true;
+  await invoke("generate_route_pages_with_handle", {
+    wikiName: $selectedWiki.name,
+  }).then((response: any) => {
+    loading = false;
+    toastStore.trigger({
+      message: response || "Route Pages generated",
+      timeout: 5000,
+      hoverable: true,
+      background: "variant-filled-success",
+    });
+  });
+}
 </script>
 
 <BaseModal bind:open={newRouteModalOpen}>
@@ -151,69 +167,88 @@ async function updatePosition() {
   </div>
 </BaseModal>
 
-<div class="flex flex-row gap-3">
-  <Button
-    class="w-40"
-    title="Create New Route"
-    onClick={() => (newRouteModalOpen = true)}
-  />
-  <Button
-    class="w-48"
-    title="Modify Encounter Types"
-    onClick={() => (encounterTypeModalOpen = true)}
-  />
-</div>
+<TabGroup>
+  <Tab bind:group={tabSet} name="routes" value={0}>Routes</Tab>
+  <Tab bind:group={tabSet} name="routes" value={1}>Generation</Tab>
+  <svelte:fragment slot="panel">
+    {#if tabSet === 0}
+      <div class="flex flex-row gap-3">
+        <Button
+          class="w-40"
+          title="Create New Route"
+          onClick={() => (newRouteModalOpen = true)}
+        />
+        <Button
+          class="w-48"
+          title="Modify Encounter Types"
+          onClick={() => (encounterTypeModalOpen = true)}
+        />
+      </div>
 
-<div class="mt-6 grid grid-cols-5 gap-x-4 gap-y-1">
-  {#each Object.keys($routes.routes) as routeName, index}
-    <div
-      class="card flex flex-row items-center justify-between !bg-transparent p-3 shadow-sm"
-    >
-      {#if routeName === routeBeingEdited}
-        <TextInput
-          bind:value={newRouteName}
-          onKeyDownHandler={(e) => {
+      <div class="mt-6 grid grid-cols-5 gap-x-4 gap-y-1">
+        {#each Object.keys($routes.routes) as routeName, index}
+          <div
+            class="card flex flex-row items-center justify-between !bg-transparent p-3 shadow-sm"
+          >
+            {#if routeName === routeBeingEdited}
+              <TextInput
+                bind:value={newRouteName}
+                onKeyDownHandler={(e) => {
             if (e.key === "Enter") {
               renameRoute(routeBeingEdited, newRouteName);
               routeBeingEdited = "";
             }
           }}
-        />
-      {:else}
-        <a href="/game-routes/{routeName}" class="w-full hover:cursor-pointer">
-          {routeName}
-        </a>
-      {/if}
-      <button
-        class="rounded-md p-1 hover:cursor-pointer hover:bg-gray-300"
-        use:popup={{
+              />
+            {:else}
+              <a
+                href="/game-routes/{routeName}"
+                class="w-full hover:cursor-pointer"
+              >
+                {routeName}
+              </a>
+            {/if}
+            <button
+              class="rounded-md p-1 hover:cursor-pointer hover:bg-gray-300"
+              use:popup={{
           event: "click",
           target: "routeMenu-" + index,
           placement: "bottom",
         }}
-      >
-        <IconDotsVertical size={16} />
-      </button>
-    </div>
-    <div class="card w-44 grid-cols-1 p-4" data-popup="routeMenu-{index}">
-      <button
-        class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
-        on:click={() => {routeBeingEdited = routeName; newRouteName = routeName;}}
-        >Rename</button
-      >
-      <button
-        class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
-        on:click={() => duplicateRoute(routeName)}>Duplicate</button
-      >
-      <button
-        class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
-        on:click={() => deleteRoute(routeName)}>Delete</button
-      >
-      <button
-        class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
-        on:click={() => {positionModalOpen = true;routeToUpdate=routeName}}
-        >Change Position</button
-      >
-    </div>
-  {/each}
-</div>
+            >
+              <IconDotsVertical size={16} />
+            </button>
+          </div>
+          <div class="card w-44 grid-cols-1 p-4" data-popup="routeMenu-{index}">
+            <button
+              class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
+              on:click={() => {routeBeingEdited = routeName; newRouteName = routeName;}}
+              >Rename</button
+            >
+            <button
+              class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
+              on:click={() => duplicateRoute(routeName)}>Duplicate</button
+            >
+            <button
+              class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
+              on:click={() => deleteRoute(routeName)}>Delete</button
+            >
+            <button
+              class="w-full rounded-md bg-gray-100 px-3 py-1 text-start text-sm hover:cursor-pointer hover:bg-gray-300"
+              on:click={() => {positionModalOpen = true;routeToUpdate=routeName}}
+              >Change Position</button
+            >
+          </div>
+        {/each}
+      </div>
+    {/if}
+    {#if tabSet === 1}
+      <Button
+        class="w-40"
+        title="Generate Pages"
+        onClick={generateRoutePages}
+        loading={loading}
+      />
+    {/if}
+  </svelte:fragment>
+</TabGroup>
